Move around due to new directories layout
