--- charts/tensorboard/Chart.yaml
@@ -1,4 +1,0 @@
-apiVersion: v1
-description: A Helm chart for running tensorboard
-name: tensorboard
-version: 0.1.0
