+++ examples/charts/tensorboard/Chart.yaml
@@ -1,0 +1,4 @@
+apiVersion: v1
+description: A Helm chart for running tensorboard
+name: tensorboard
+version: 0.1.0
