+++ examples/charts/tensorboard/values.yaml
@@ -1,0 +1,6 @@
+# Default values for tensorboard.
+# This is a YAML-formatted file.
+# Declare variables to be passed into your templates.
+#
+# The log directory to use with TensorBoard.
+logDir:
