--- release/releaser.yaml
@@ -1,22 +1,0 @@
-apiVersion: extensions/v1beta1 # for versions before 1.6.0 use extensions/v1beta1
-kind: ReplicaSet
-metadata:
-  name: releaser
-  labels:
-    app: releaser
-spec:
-  replicas: 1
-  selector:
-    matchLabels:
-      app: releaser
-  template:
-    metadata:
-      labels:
-        app: releaser
-    spec:
-      containers:
-      - name: releaser
-        image: gcr.io/tf-on-k8s-releasing/releaser:latest
-        workingDir: /opt/tf_k8s_releaser
-        command:
-          - /opt/tf_k8s_releaser/launch_build.sh
