--- test-infra/airflow/deployment.yaml
@@ -1,78 +1,0 @@
-kind: Service
-apiVersion: v1
-metadata:
-  name: airflow
-spec:
-  selector:
-    name: airflow
-  ports:
-  - protocol: TCP
-    port: 80
-    targetPort: 8080
-
----
-
-apiVersion: extensions/v1beta1
-kind: Deployment
-metadata:
-  name: airflow
-spec:
-  replicas: 1
-  template:
-    metadata:
-      labels:
-        name: airflow
-    spec:
-      containers:
-      - name: airflow
-        image: gcr.io/mlkube-testing/airflow:v20180112-544fb73-dirty-e514e6
-        # What arguments do we need to set to use the LocalExecutor
-        command:
-          # Tells entrypoint.sh to start the webserver
-          - /entrypoint.sh
-          - webserver
-        env:
-        - name: POSTGRES_HOST
-          value: localhost
-        - name: GOOGLE_APPLICATION_CREDENTIALS
-          value: /var/secrets/google/key.json
-        # SRC_DIR is the location where we create a directory for each
-        # pipeline run. This directory is used to communicate data between
-        # steps in the pipeline.
-        - name: SRC_DIR
-          value: /var/lib/data/runs
-        volumeMounts:
-        - name: google-cloud-key
-          mountPath: /var/secrets/google
-        - name: airflow-data
-          mountPath: /var/lib/data
-        ports:
-        - containerPort: 8080
-      - name: postgres
-        image: postgres:10.1-alpine
-        env:
-        # For a list of POSTGRES environment variables see
-        # https://hub.docker.com/_/postgres/
-        - name: POSTGRES_USER
-          value: airflow
-        - name: POSTGRES_PASSWORD
-          value: airflow
-        - name: POSTGRES_DB
-          value: airflow
-        # PGDATA controls where the data for the DB is stored. We want to
-        # store it on our PD.
-        - name: PGDATA
-          value: /var/lib/postgresql/data/airflow-postgre
-        ports:
-        - containerPort: 5432
-        volumeMounts:
-        - name: airflow-data
-          mountPath: /var/lib/postgresql/data
-      volumes:
-        - name: airflow-data
-          gcePersistentDisk:
-            pdName: airflow-data
-            fsType: ext4
-        - name: google-cloud-key
-          secret:
-            secretName: airflow-key
