--- test-infra/airflow/entrypoint.sh
@@ -1,63 +1,0 @@
-#!/usr/bin/env bash
-#
-# A startup script for Airflow.
-#
-# Based on
-# https://raw.githubusercontent.com/puckel/docker-airflow/master/script/entrypoint.sh
-
-AIRFLOW_HOME="/usr/local/airflow"
-CMD="airflow"
-TRY_LOOP="20"
-
-# TODO(jlewi): This is currently getting overwritten in the YAML file.
-: ${POSTGRES_HOST:="postgres"}
-: ${POSTGRES_PORT:="5432"}
-: ${POSTGRES_USER:="airflow"}
-: ${POSTGRES_PASSWORD:="airflow"}
-: ${POSTGRES_DB:="airflow"}
-
-# TODO(jlewi): Should we make the key into the Docker image rather than doing
-# it on startup?
-: ${FERNET_KEY:=$(python -c "from cryptography.fernet import Fernet; FERNET_KEY = Fernet.generate_key().decode(); print(FERNET_KEY)")}
-
-
-# Update airflow config - Fernet key
-sed -i "s|\$FERNET_KEY|$FERNET_KEY|" "$AIRFLOW_HOME"/airflow.cfg
-
-# Wait for Postresql
-# TODO(jlewi): If we are just using LocalExecutor we should always be starting
-# the webserver so we could maybe eliminate the if statement and simplify
-# this.
-if [ "$1" = "webserver" ] || [ "$1" = "worker" ] || [ "$1" = "scheduler" ] ; then
-  i=0
-  while ! nc -z $POSTGRES_HOST $POSTGRES_PORT >/dev/null 2>&1 < /dev/null; do
-    i=$((i+1))
-    if [ "$1" = "webserver" ]; then
-      echo "$(date) - waiting for ${POSTGRES_HOST}:${POSTGRES_PORT}... $i/$TRY_LOOP"
-      if [ $i -ge $TRY_LOOP ]; then
-        echo "$(date) - ${POSTGRES_HOST}:${POSTGRES_PORT} still not reachable, giving up"
-        exit 1
-      fi
-    fi
-    sleep 10
-  done
-fi
-
-if [ -n "${GOOGLE_APPLICATION_CREDENTIALS}" ]; then
-  echo GOOGLE_APPLICATION_CREDENTIALS=${GOOGLE_APPLICATION_CREDENTIALS}
-  # Configure gcloud to use this service account
-  gcloud auth activate-service-account --key-file=${GOOGLE_APPLICATION_CREDENTIALS}
-fi
-
-# TODO(jlewi): Can we get get rid of the sed and just put this into our config
-# file?
-sed -i "s#sql_alchemy_conn = postgresql+psycopg2://airflow:airflow@postgres/airflow#sql_alchemy_conn = postgresql+psycopg2://$POSTGRES_USER:$POSTGRES_PASSWORD@$POSTGRES_HOST:$POSTGRES_PORT/$POSTGRES_DB#" "$AIRFLOW_HOME"/airflow.cfg
-sed -i "s#broker_url = redis://redis:6379/1#broker_url = redis://$REDIS_PREFIX$REDIS_HOST:$REDIS_PORT/1#" "$AIRFLOW_HOME"/airflow.cfg
-echo "Initialize database..."
-$CMD initdb
-echo start the webserver
-exec $CMD webserver &
-# TODO(jlewi): How do we capture logs from the scheduler? Maybe we should
-# move it into its own container?
-echo start the scheduler
-exec $CMD scheduler
