+++ test/test-infra/airflow/Dockerfile.prow.template
@@ -1,0 +1,20 @@
+# Copyright 2017 The Kubernetes Authors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+# This Dockerfile is used by ProwJobs to launch our Airflow jobs.
+#
+# TODO(jlewi) With Docker v17.0 we should just be able to use Docker
+# build args to specify the base image.
+FROM {{base_image}}
+ENTRYPOINT ["python", "-m", "py.airflow"]
